First experimental setups
First experiments with models and their characteristics. Hubs of different types of producers and consumers, a first consumer profile (Wicon office) has also been deposited which works well. The resolution of the data is balanced in 15 minute intervals and at 1000kwh over the year. Therefore the profile can be adapted to different strong consumers in the office area. On the generator side a generator with fixed speed is used.